make user confirmable

+++ db/migrate/20230801113821_create_foods.rb
@@ -4,7 +4,7 @@ class CreateFoods < ActiveRecord::Migration[7.0]
       t.string :name
       t.string :measurement_unit
       t.float :price
-      t.float :quantity
+      t.integer :quantity
       t.references :user, null: false, foreign_key: true, index: true
 
       t.timestamps
